requirements.txt
aiofiles==23.2.1
anyio==4.2.0
certifi==2023.11.17
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
PyYAML==6.0.1
sniffio==1.3.0
yadisk==2.1.0
